Extract Card Ranks

`card_rank()` retrieves the rank field from a vctrs “card” record (class "card"). When applied to a plain character vector, it returns the vector unchanged. This function was specifically designed to work with vctrs “card” objects created by create_shuffled_deck().

Usage

card_rank(x, ...)

Arguments

x

Either:

A vctrs “card” record (class "card") as returned by create_shuffled_deck(), or any other "card" object with a rank field; or

A character vector of card ranks (e.g.\ "Two", "Jack", "Queen").

If x is a "card" record, returns a character vector of the rank field; if x is a character vector, returns x unchanged.

Value

Depending on the method:

card

A character vector (same length as x), containing the rank field of each card.

character

The input character vector, returned as it is.

Details

- For objects of class "card", card_rank.card(x) extracts the rank field via vctrs::field(x, "rank"), returning a character vector of card ranks (e.g.\ "Ace").

- If x is a character vector, card_rank.character(x) simply returns x unchanged.

- For any other class, card_rank.default(x) throws an error indicating that only "card" or character inputs are accepted.

Examples

deck <- create_shuffled_deck()
card_rank(deck)
#>   [1] "Ace"   "Five"  "Eight" "Nine"  "Ace"   "Five"  "Three" "Three" "Nine" 
#>  [10] "Queen" "Queen" "Jack"  "Nine"  "Three" "Nine"  "Seven" "Four"  "Seven"
#>  [19] "Ten"   "Ace"   "Seven" "Six"   "Ten"   "Two"   "King"  "Three" "Ten"  
#>  [28] "Four"  "Three" "Five"  "Ace"   "Four"  "Four"  "Queen" "King"  "Two"  
#>  [37] "Ten"   "Nine"  "Six"   "Eight" "Jack"  "Queen" "Jack"  "Jack"  "King" 
#>  [46] "Seven" "Eight" "Seven" "Two"   "King"  "Six"   "Five"  "Eight" "Four" 
#>  [55] "Six"   "Four"  "Ten"   "Jack"  "Nine"  "Three" "Three" "Seven" "Seven"
#>  [64] "Six"   "Seven" "Two"   "Nine"  "Eight" "Jack"  "Six"   "Ten"   "Ten"  
#>  [73] "Six"   "Three" "Seven" "Eight" "Jack"  "Nine"  "Ten"   "Seven" "Ace"  
#>  [82] "Jack"  "Seven" "Three" "King"  "Jack"  "Queen" "Ten"   "Eight" "Jack" 
#>  [91] "Six"   "Queen" "Jack"  "Two"   "Five"  "Eight" "Queen" "Eight" "Jack" 
#> [100] "Three" "King"  "Five"  "Ace"   "King"  "Five"  "Ace"   "Ace"   "Queen"
#> [109] "Five"  "King"  "Six"   "Eight" "King"  "Five"  "Two"   "Ten"   "Five" 
#> [118] "King"  "King"  "Four"  "Six"   "Nine"  "Ace"   "Ace"   "Queen" "Two"  
#> [127] "Nine"  "Five"  "Jack"  "Two"   "Two"   "Three" "Four"  "Ace"   "Eight"
#> [136] "Ace"   "Three" "Two"   "Ten"   "Eight" "Ten"   "Queen" "King"  "Queen"
#> [145] "Ten"   "Nine"  "Seven" "Two"   "Five"  "King"  "Seven" "King"  "Queen"
#> [154] "Three" "Four"  "Ten"   "Jack"  "Nine"  "Two"   "Ace"   "Six"   "Jack" 
#> [163] "Queen" "Two"   "Ace"   "Four"  "Six"   "Five"  "Ace"   "Four"  "King" 
#> [172] "King"  "Nine"  "Five"  "Two"   "Two"   "Three" "Six"   "Five"  "Four" 
#> [181] "Queen" "Four"  "Ace"   "Three" "Four"  "Eight" "Six"   "Nine"  "Queen"
#> [190] "Nine"  "Seven" "Queen" "Four"  "Three" "Ten"   "Jack"  "Eight" "Ten"  
#> [199] "Nine"  "Six"   "Five"  "Two"   "Six"   "Eight" "Seven" "Seven" "Eight"
#> [208] "Four"