Create a Shuffled Blackjack Shoe

Creates a shuffled Blackjack shoe consisting of multiple standard 52-card decks combined. Each standard deck includes cards from all four suits ("♠", "♥", "♦", "♣") and 13 ranks (2–10, J, Q, K, A). The function has been updated to include vctrs class "card" record with fields: name, rank, suit_symbol, suit, and is_face. name is basically the card's name, rank is the English word for the rank, suit_symbol is the Unicode symbol for the suit, suit is the English word for the suit, and is_face indicates if the card is a face card (Jack, Queen, King).

Usage

create_shuffled_deck(noOfDecks = 4)

Arguments

noOfDecks

Integer (≥1). Number of decks to include in the shoe (default = 4). A typical Blackjack shoe contains 4 to 8 decks.

Value

A [vctrs][vctrs::new_rcrd] “card” record vector of length `52 * noOfDecks`. Each element is an S3 object of class `"card"` with these fields:

name – character, e.g. "2♣", "A♠"

rank – character, one of "Two", "Three", …, "Ace"

suit_symbol– character, one of "♠", "♥", "♦", "♣"

suit – character, one of "Spade", "Heart", "Diamond", "Club"

is_face – logical; TRUE for Jack/Queen/King, FALSE otherwise

After update, the creating_shuffle_deck now able to return a vctrs record with the above fields instead of just a character vector.

Details

Because we use vctrs::new_rcrd(...), the returned object is a vctrs record with class card. You can extract the underlying character names with as.character(), or access fields via vctrs::field(). The print.card() method will display all card names, and format.card() returns the card’s "name" field.”

Examples

deck <- create_shuffled_deck()
length(deck)                   # Should be 208 (4 decks of 52 as default setting)
#> [1] 208
deck                           # Prints the deck of cards
#> card[208]
#> 3♥ A♦ 5♦ 4♦ 4♠ 6♠ 6♦ Q♠ 9♦ A♣ Q♥ 10♥ 2♥ 9♦ 3♦ 8♣ 10♠ A♠ 8♣ 7♦ Q♦ 2♠ A♦ A♠ 10♥ A♣ 7♥ J♦ 4♦ 8♥ 4♦ A♠ J♣ 9♥ 2♥ 9♣ 4♦ 2♣ 7♠ 4♠ J♠ K♦ 9♦ 2♦ 4♥ 4♥ 5♣ 5♠ 4♣ 6♦ 6♣ 2♦ 9♣ 8♠ 10♠ 9♦ Q♣ 8♥ A♥ 9♠ 5♠ J♥ 2♦ 6♠ 8♠ 6♠ 2♣ 7♦ 4♣ K♥ 5♣ Q♣ 6♦ K♣ K♠ J♠ Q♦ J♣ 8♦ 10♥ 3♣ J♦ 9♠ Q♥ J♠ 10♠ A♥ 7♣ Q♥ 9♥ 8♠ 4♠ 4♣ 7♣ Q♠ K♣ 3♥ K♦ J♥ A♦ 5♥ 5♥ 10♦ 3♦ 10♣ 8♦ Q♥ K♦ 10♦ 9♠ A♠ 7♥ 4♥ 3♠ 9♥ J♦ K♥ A♦ 3♦ 5♣ Q♣ 8♦ J♥ J♣ 5♦ 8♥ 6♥ 5♠ 6♦ 2♣ 4♠ 3♥ 2♠ J♥ 7♣ K♥ 10♦ 5♦ 9♠ A♥ 2♠ 6♣ K♦ 8♦ Q♦ 10♠ K♠ K♠ K♣ 2♥ 3♣ 8♠ 10♣ 2♣ 10♥ 9♣ J♣ 3♠ 4♥ K♥ 8♣ J♠ 3♣ 3♥ 6♣ 10♣ 6♣ 6♥ 2♥ 7♥ 7♠ Q♠ 10♣ 7♠ 6♠ Q♦ 8♥ A♣ 5♥ 6♥ 4♣ 3♠ 7♥ Q♣ 2♦ 9♣ 8♣ 5♣ 3♣ 3♦ 5♠ 10♦ 5♥ 2♠ 7♦ 9♥ J♦ K♠ K♣ 6♥ 5♦ Q♠ 3♠ 7♦ A♥ A♣ 7♣ 7♠ 
deck2 <- create_shuffled_deck(noOfDecks = 6)
length(deck2)                 # Should be 312 (6 * 52)
#> [1] 312