Player Turn Interaction

Handles the player's turn in a Blackjack game, allowing them to hit or stand.

Usage

player_turn(player_hand, deck, input_fn = readline)

Arguments

player_hand

A vector of cards representing the player's hand.

deck

A vector of cards representing the remaining deck.

input_fn

a holder to obtain player actions (defaults to readline); can be overridden to inject fake input for unit tests or if the user want to add fake actions.

Value

A list containing:

hand

The updated player hand after the turn.

deck

The updated deck after the player's turn.

total

The total value of the player's hand.

Details

1. Displaying the hand:

Before each prompt, the function prints the player’s current hand and its total value as computed by card_value(player_hand). Card names are shown via format(player_hand) while also account for vctrs class.

2. Hit or Stand:

If the user types "hit" or "h", the function calls deal_hand(deck, 1) to draw one card. That card is appended to player_hand, and deck is updated accordingly. If the user types "stand" or "s", the loop ends and the final total is printed. Any other input triggers an “Invalid choice” message and re-prompts.

3. Stopping conditions: The loop continues until the player chooses to stand or their total reaches 21 or higher. When player reach 21 then the loop end immediately.

Examples

# Example: Player always stands immediately
deck <- create_shuffled_deck()
player_hand <- deal_hand(deck, 2)$hand
fake_input <- function(prompt) "stand"
player_turn(player_hand, deck, input_fn = fake_input)
#> [1] Your hand: 5♦ 4♠ (Total = 9 )
#> [1] You stand at 9
#> $hand
#> card[2]
#> 5♦ 4♠ 
#> 
#> $deck
#> card[208]
#> 5♦ 4♠ 6♠ 3♣ 10♣ 9♥ 9♠ 7♣ 8♠ 5♣ K♥ 8♣ K♠ K♦ A♣ 5♣ 4♠ 3♦ 10♥ 4♥ A♠ K♠ 10♣ J♥ 6♠ J♠ 7♦ 8♦ J♥ 10♠ 10♥ 10♦ Q♦ Q♦ 7♥ 8♦ 2♣ 7♥ 7♥ 8♠ A♦ Q♠ K♠ 6♦ 4♦ 7♦ J♦ 3♦ 6♥ K♣ A♣ 8♦ 5♠ 9♣ K♥ 4♠ J♦ 2♦ 2♣ 7♦ 5♣ 8♠ Q♦ 9♦ 10♠ 4♦ 8♣ 2♥ 6♣ 4♦ 7♣ 5♦ 4♣ Q♥ A♠ 8♥ Q♠ 5♣ 6♥ 3♣ 7♦ K♥ 6♥ 8♠ 6♦ 5♦ 4♥ J♣ Q♥ 3♦ 3♠ J♣ K♦ 4♦ 9♠ J♠ 6♣ Q♠ 2♣ Q♥ 2♣ 9♦ 6♦ 9♦ Q♠ 3♥ 6♠ 4♥ 10♠ J♦ 3♦ 7♥ 9♦ 9♥ 9♥ 4♠ K♦ 3♥ A♦ A♥ 2♥ 2♠ J♠ A♥ 3♠ 6♥ J♣ 3♣ K♣ Q♦ 2♠ A♥ 6♣ 5♦ 5♠ A♦ 10♠ J♦ 2♥ 8♥ 10♦ 9♠ 5♠ 8♣ 10♥ 10♦ A♣ K♠ A♦ 7♠ 8♣ 3♠ 3♣ Q♣ K♦ 9♥ Q♥ 8♦ 2♥ 5♥ J♥ 8♥ 5♥ Q♣ 4♣ 5♥ J♥ 10♦ 5♥ 7♠ 9♠ J♣ 4♣ 2♦ Q♣ K♣ A♣ 6♦ 2♠ 2♦ 10♣ 10♥ 2♠ 5♠ 9♣ 4♣ 7♠ 8♥ 4♥ 3♥ A♥ 6♣ 7♠ 6♠ J♠ 3♥ 9♣ A♠ 7♣ Q♣ 9♣ K♥ K♣ 10♣ 2♦ A♠ 7♣ 3♠ 
#> 
#> $total
#> [1] 9
#>