Extract Card Names

`card_name()` retrieves the name field from a vctrs “card” record (class "card"). When applied to a plain character vector, it returns the vector unchanged. This function was specifically designed to work with vctrs “card” objects created by create_shuffled_deck().

Usage

card_name(x, ...)

Arguments

x

Either:

A vctrs “card” record (class "card") as returned by create_shuffled_deck(), or any other "card" object with a name field; or

A character vector of card names (e.g.\ "A♠", "10♦", "K♥").

If x is a "card" record, returns a character vector of the name field; if x is a character vector, returns x unchanged.

Value

Depending on the method:

card

A character vector (same length as x), containing the name field of each card.

character

The input character vector, returned as it is.

Details

- For objects of class "card", card_name.card(x) extracts the name field via vctrs::field(x, "name"), returning a character vector of card names (e.g.\ "A♠").

- If x is a character vector, card_name.character(x) simply returns x unchanged.

- For any other class, card_name.default(x) throws an error indicating that only "card" or character inputs are accepted.

Examples

deck <- create_shuffled_deck()
card_name(deck)
#>   [1] "J♣"  "5♦"  "K♣"  "10♠" "K♠"  "6♣"  "K♦"  "3♥"  "6♥"  "A♦"  "2♣"  "5♠" 
#>  [13] "8♥"  "Q♥"  "Q♠"  "K♦"  "2♠"  "8♥"  "A♥"  "10♠" "2♣"  "Q♣"  "10♣" "6♠" 
#>  [25] "4♦"  "K♥"  "7♠"  "10♦" "6♦"  "6♦"  "Q♣"  "2♦"  "3♣"  "Q♠"  "8♦"  "10♣"
#>  [37] "4♠"  "4♣"  "A♦"  "10♣" "5♥"  "10♥" "J♦"  "J♥"  "A♣"  "6♣"  "5♠"  "6♦" 
#>  [49] "J♥"  "8♣"  "9♥"  "4♦"  "8♣"  "3♦"  "7♦"  "J♣"  "7♣"  "6♥"  "3♣"  "8♠" 
#>  [61] "6♦"  "7♠"  "4♦"  "K♠"  "9♠"  "10♦" "7♦"  "J♥"  "8♣"  "5♣"  "7♦"  "2♥" 
#>  [73] "2♦"  "3♦"  "K♣"  "7♣"  "2♠"  "K♠"  "5♦"  "5♥"  "K♥"  "K♦"  "6♣"  "8♦" 
#>  [85] "K♠"  "Q♠"  "9♦"  "A♠"  "9♣"  "3♠"  "4♣"  "3♣"  "3♠"  "7♠"  "10♠" "J♦" 
#>  [97] "J♠"  "10♦" "Q♥"  "J♠"  "10♥" "9♠"  "2♦"  "9♦"  "8♦"  "7♥"  "4♠"  "9♠" 
#> [109] "6♣"  "A♠"  "K♣"  "Q♦"  "2♥"  "9♥"  "A♣"  "8♥"  "5♣"  "7♣"  "7♠"  "4♠" 
#> [121] "6♥"  "5♥"  "A♠"  "J♠"  "6♠"  "8♠"  "4♥"  "A♠"  "3♣"  "J♠"  "9♣"  "10♦"
#> [133] "8♠"  "7♦"  "2♠"  "8♥"  "4♣"  "2♠"  "A♦"  "2♥"  "K♥"  "3♦"  "3♥"  "4♣" 
#> [145] "7♥"  "4♥"  "Q♦"  "5♥"  "K♦"  "K♣"  "J♦"  "Q♠"  "4♥"  "3♥"  "6♠"  "K♥" 
#> [157] "9♠"  "7♥"  "J♦"  "A♥"  "2♦"  "10♠" "Q♥"  "5♦"  "9♥"  "10♣" "J♣"  "2♥" 
#> [169] "6♠"  "A♣"  "Q♦"  "Q♥"  "A♣"  "2♣"  "9♦"  "3♥"  "8♦"  "10♥" "J♥"  "2♣" 
#> [181] "5♠"  "Q♦"  "9♣"  "A♥"  "5♠"  "5♣"  "10♥" "4♠"  "7♥"  "4♥"  "9♦"  "5♦" 
#> [193] "6♥"  "Q♣"  "3♠"  "A♥"  "4♦"  "8♠"  "5♣"  "9♣"  "A♦"  "7♣"  "3♠"  "9♥" 
#> [205] "3♦"  "J♣"  "Q♣"  "8♣"