Deal a Blackjack Hand

Deals cards from a given deck and removes them from the deck using C++ for performance. If the deck fall below 25

Usage

deal_hand(deck, n = 2)

Arguments

deck

A deck of cards, typically the output of create_shuffled_deck(). Internally, this should be a vctrs record with class "card" (each element has fields name, rank, suit_symbol, suit, is_face) or a plain character vector of card names.

n

Number of cards to deal (default = 2).

Value

A list with:

hand

Character vector or "card" record of dealt cards depend on the deck used.

deck

The updated deck after removing dealt cards.

Details

After the update, deal_hand() is now an S3 generic with methods for both card records and the old character. If the user use this deal_hand() with a character vector, it will still work as before, however, it lack the ability to check with create_shuffle_deck for whether the amount of cards fall below 25 If deck inherits from class "card", it dispatch to deal_hand.card(). If deck is a bare character vector, it dispatch to deal_hand.character(). If is an unexpected class then it dispatch to deal_hand.default() which throws an error.

Examples

# Create a shuffled deck and deal a hand of 2 cards
deck <- create_shuffled_deck()
result <- deal_hand(deck, 2)
result$hand
#> card[2]
#> 4♠ 10♦ 
length(result$deck)
#> [1] 206
# Automatic reshuffle
small_deck <- head(create_shuffled_deck(), 6)
deal_hand(small_deck, 2)
#> creating new deck...
#> $hand
#> card[2]
#> 3♦ 5♠ 
#> 
#> $deck
#> card[206]
#> 2♠ J♠ J♣ 10♠ J♥ 6♦ 10♣ 7♣ 4♣ 9♠ Q♠ K♦ 4♥ Q♦ 5♦ 6♥ 4♠ 3♥ 6♦ 5♥ 3♠ 8♠ A♥ K♥ 5♠ J♦ 3♠ 8♠ 7♠ J♥ 10♠ 7♦ 6♥ 9♥ 2♦ A♣ 7♣ 5♣ 4♣ A♦ 9♣ 9♥ 9♦ 7♥ 10♦ 3♦ 6♦ A♣ 8♦ A♠ 8♣ J♠ Q♥ 7♥ Q♦ 4♦ 8♦ 10♥ 3♣ 3♥ 10♠ 7♥ 4♦ 6♣ 6♦ 4♠ 6♠ 4♥ 6♥ 5♠ 2♠ 6♣ 3♥ K♣ J♦ J♥ 10♦ Q♦ J♣ A♦ K♠ 10♦ K♦ K♥ J♦ K♥ 3♥ 2♣ 10♥ A♥ 10♣ A♥ K♦ Q♦ 4♥ 2♣ 9♦ 3♣ 5♦ 10♠ 2♦ 5♣ J♣ 5♦ 7♠ 3♠ Q♣ K♣ K♥ K♣ 4♦ 3♣ 9♠ 2♦ 5♦ Q♠ 2♠ 6♠ A♠ 9♦ 8♣ 6♣ 7♣ J♦ 5♥ K♠ 10♥ 2♠ Q♣ K♠ 4♠ 5♣ 8♣ 9♣ 2♥ 9♣ A♥ 9♠ 4♠ 8♥ 8♦ 2♥ 3♠ 4♣ A♣ 6♣ J♠ 5♠ 8♥ 10♥ Q♠ A♣ 7♠ 6♠ 5♥ 6♥ 8♥ 4♦ 2♥ A♦ 9♥ 9♦ K♠ 9♠ 10♦ 7♥ 7♠ Q♠ A♦ 2♣ 9♣ 4♣ 8♥ 5♥ A♠ 2♦ 2♥ 3♦ K♣ 3♣ 7♦ 7♦ 8♠ J♥ Q♥ 9♥ 3♦ Q♣ Q♥ J♠ 8♣ 10♣ 2♣ 7♣ 10♣ 4♥ 7♦ Q♥ 5♣ 8♠ 8♦ 6♠ J♣ K♦ A♠ Q♣ 
#>